pull out getBestNextGuess
