+++ src/components/Game.tsx
@@ -2,9 +2,8 @@ import React, { useCallback } from "react";
 import { useAppDispatch, useAppSelector } from "../store";
 import Victory from "./Victory";
 import gameSlice from "../store/gameSlice";
-import { datamuseApi } from "../services/datamuse-api";
 import Guess from "./Guess";
-import { chooseBestGuess } from "../utilities/guess";
+import { getBestNextGuess } from "../utilities/guess";
 import WordResult from "./WordResult";
 
 const EMPTY: [] = [];
@@ -31,40 +30,17 @@ const Game: React.FC<{
     [dispatch, answer]
   );
 
-  const autoGuess = useCallback(() => {
-    const spelledLike = new Array(5)
-      .fill(null)
-      .map((_, i) => {
-        const correctLetter = guesses.find(
-          (guess) => guess[i].inPosition && guess[i].inWord
-        );
-        return correctLetter?.[i].character ?? "?";
-      })
-      .join("");
-
-    return datamuseApi
-      .getWordsInfo({
-        spelledLike: spelledLike,
-        metadata: {
-          frequency: true,
-        },
-      })
-      .then((wordsInfo) => {
-        const pastGuesses = guesses.map((guess) =>
-          guess.map((status) => status.character).join("")
-        );
-
-        const bestGuess = chooseBestGuess({
-          wordsInfo,
-          pastGuesses,
-          eliminatedLetters,
-          foundLetters,
-        });
-
-        if (bestGuess) {
-          handleSubmitGuess(bestGuess);
-        }
-      });
+  const autoGuess = useCallback(async () => {
+    const bestNextGuess = await getBestNextGuess(
+      guesses,
+      eliminatedLetters,
+      foundLetters
+    );
+    if (bestNextGuess) {
+      handleSubmitGuess(bestNextGuess);
+    } else {
+      console.error("No best next guess found");
+    }
   }, [guesses, eliminatedLetters, foundLetters, handleSubmitGuess]);
 
   return (
